Formule added and pyramid percentage corrected
The formulae of the population indicators were added. Some explanatory comments were added in the pdf file.
The percentage values of the population pyramid were wrong. They were corrected and the pyramids were updated with the corrected population values. The denominator of the population ratio in each sex-age group should be the total population of the same year and not the sex-specific total population of that year.